add second version of worley noise
